requirements.txt
dnspython
pandas
python-dotenv
openpyxl
requests